requirements.txt
numpy
Pillow
pyhackrf
matplotlib